Split into components and create separate csv files
