+++ src/components/App.css
@@ -12,7 +12,3 @@
 .App-title {
   font-size: 1.5em;
 }
-
-.App-intro {
-  font-size: large;
-}
